# Prefer strings.Builder over bytes.Buffer for building the bulk body.