The strings.Builder type is designed to mimimize memory copying.  

**Note strings.Builder available since go 1.10**