@rwynn 2019-11-03
For anyone interested I tried a simple memory profile on the following test code..

```golang
func TestBulkAlloc(t *testing.T) {
        json := "{\"foo\": 1}"
        client, _ := elastic.NewClient()
        bulk := elastic.NewBulkService(client)
        for j := 0; j < 100; j++ {
                for i := 0; i < 100000; i++ {
                        req := elastic.NewBulkIndexRequest()
                        req.Index("test")
                        req.Doc(json)
                        bulk.Add(req)
                }
                bulk.Do(context.Background())
                bulk.Reset()
        }
}
```

Ran the test and analysis using

```
go test -memprofile alloc.prof -bench .
go tool pprof alloc.prof
```

This is the result with 7.0.8

```
(pprof) list bodyAsString
Total: 22.35GB
ROUTINE ======================== github.com/olivere/elastic/v7.(*BulkService).bodyAsString in /home/vagrant/go/pkg/mod/github.com/olivere/elastic/v7@v7.0.8/bulk.go
  372.87MB     7.40GB (flat, cum) 33.09% of Total
         .          .    209:   return len(s.requests)
         .          .    210:}
         .          .    211:
         .          .    212:func (s *BulkService) bodyAsString() (string, error) {
         .          .    213:   // Pre-allocate to reduce allocs
  372.87MB     7.03GB    214:   buf := bytes.NewBuffer(make([]byte, 0, s.EstimatedSizeInBytes()))
         .          .    215:
         .          .    216:   for _, req := range s.requests {
         .          .    217:           source, err := req.Source()
         .          .    218:           if err != nil {
         .          .    219:                   return "", err
         .          .    220:           }
         .          .    221:           for _, line := range source {
         .          .    222:                   buf.WriteString(line)
         .          .    223:                   buf.WriteByte('\n')
         .          .    224:           }
         .          .    225:   }
         .          .    226:
         .   372.87MB    227:   return buf.String(), nil
         .          .    228:}
         .          .    229:
         .          .    230:// Do sends the batched requests to Elasticsearch. Note that, when successful,
         .          .    231:// you can reuse the BulkService for the next batch as the list of bulk
         .          .    232:// requests is cleared on success.
```


And with this PR applied

```
(pprof) list bodyAsString
Total: 21.84GB
ROUTINE ======================== github.com/olivere/elastic/v7.(*BulkService).bodyAsString in /vagrant/elastic/bulk.go
         0     6.98GB (flat, cum) 31.97% of Total
         .          .    209:}
         .          .    210:
         .          .    211:func (s *BulkService) bodyAsString() (string, error) {
         .          .    212:   // Pre-allocate to reduce allocs
         .          .    213:   var buf strings.Builder
         .     6.98GB    214:   buf.Grow(int(s.EstimatedSizeInBytes()))
         .          .    215:
         .          .    216:   for _, req := range s.requests {
         .          .    217:           source, err := req.Source()
         .          .    218:           if err != nil {
         .          .    219:                   return "", err
```

So, I think this PR does improve allocations a bit. Specifically, at buf.String() where the PR code doesn't allocate.