@olivere 2019-11-04
I ran [the existing benchmarks](https://github.com/olivere/elastic/blob/release-branch.v6/bulk_test.go#L558-L631) and benchcmp'd old and new.

Here are the results for v6:

```
benchmark                                              old ns/op      new ns/op      delta
BenchmarkBulkDeleteRequestSerialization/stdlib-8       1260           1258           -0.16%
BenchmarkBulkDeleteRequestSerialization/easyjson-8     533            551            +3.38%
BenchmarkBulkIndexRequestSerialization/stdlib-8        2205           2178           -1.22%
BenchmarkBulkIndexRequestSerialization/easyjson-8      1466           1468           +0.14%
BenchmarkBulkEstimatedSizeInBytesWith1Request-8        7589           7224           -4.81%
BenchmarkBulkEstimatedSizeInBytesWith100Requests-8     743534         719697         -3.21%
BenchmarkBulkAllocs/1000_docs_with_64_byte-8           51916          32417          -37.56%
BenchmarkBulkAllocs/1000_docs_with_1_KiB-8             280967         192871         -31.35%
BenchmarkBulkAllocs/1000_docs_with_4_KiB-8             1106163        898677         -18.76%
BenchmarkBulkAllocs/1000_docs_with_16_KiB-8            5783590        3959724        -31.54%
BenchmarkBulkAllocs/1000_docs_with_64_KiB-8            19735519       10553467       -46.53%
BenchmarkBulkAllocs/1000_docs_with_256_KiB-8           80118663       68952867       -13.94%
BenchmarkBulkAllocs/1000_docs_with_1_MiB-8             2400098864     1791674468     -25.35%
BenchmarkBulkUpdateRequestSerialization/stdlib-8       2347           2510           +6.95%
BenchmarkBulkUpdateRequestSerialization/easyjson-8     1075           1083           +0.74%

benchmark                                              old allocs     new allocs     delta
BenchmarkBulkDeleteRequestSerialization/stdlib-8       9              9              +0.00%
BenchmarkBulkDeleteRequestSerialization/easyjson-8     5              5              +0.00%
BenchmarkBulkIndexRequestSerialization/stdlib-8        15             15             +0.00%
BenchmarkBulkIndexRequestSerialization/easyjson-8      11             11             +0.00%
BenchmarkBulkEstimatedSizeInBytesWith1Request-8        47             47             +0.00%
BenchmarkBulkEstimatedSizeInBytesWith100Requests-8     4412           4412           +0.00%
BenchmarkBulkAllocs/1000_docs_with_64_byte-8           2              1              -50.00%
BenchmarkBulkAllocs/1000_docs_with_1_KiB-8             5              3              -40.00%
BenchmarkBulkAllocs/1000_docs_with_4_KiB-8             17             12             -29.41%
BenchmarkBulkAllocs/1000_docs_with_16_KiB-8            73             50             -31.51%
BenchmarkBulkAllocs/1000_docs_with_64_KiB-8            270            148            -45.19%
BenchmarkBulkAllocs/1000_docs_with_256_KiB-8           1157           1002           -13.40%
BenchmarkBulkAllocs/1000_docs_with_1_MiB-8             15027          15025          -0.01%
BenchmarkBulkUpdateRequestSerialization/stdlib-8       19             19             +0.00%
BenchmarkBulkUpdateRequestSerialization/easyjson-8     10             10             +0.00%

benchmark                                              old bytes      new bytes      delta
BenchmarkBulkDeleteRequestSerialization/stdlib-8       1560           1560           +0.00%
BenchmarkBulkDeleteRequestSerialization/easyjson-8     1408           1408           +0.00%
BenchmarkBulkIndexRequestSerialization/stdlib-8        1192           1192           +0.00%
BenchmarkBulkIndexRequestSerialization/easyjson-8      1040           1040           +0.00%
BenchmarkBulkEstimatedSizeInBytesWith1Request-8        5368           5367           -0.02%
BenchmarkBulkEstimatedSizeInBytesWith100Requests-8     542000         542006         +0.00%
BenchmarkBulkAllocs/1000_docs_with_64_byte-8           262199         131107         -50.00%
BenchmarkBulkAllocs/1000_docs_with_1_KiB-8             2180221        1090248        -49.99%
BenchmarkBulkAllocs/1000_docs_with_4_KiB-8             8338490        4173188        -49.95%
BenchmarkBulkAllocs/1000_docs_with_16_KiB-8            33156584       16628393       -49.85%
BenchmarkBulkAllocs/1000_docs_with_64_KiB-8            135027777      67702827       -49.86%
BenchmarkBulkAllocs/1000_docs_with_256_KiB-8           586296036      315836237      -46.13%
BenchmarkBulkAllocs/1000_docs_with_1_MiB-8             5262552840     4214967880     -19.91%
BenchmarkBulkUpdateRequestSerialization/stdlib-8       2272           2272           +0.00%
BenchmarkBulkUpdateRequestSerialization/easyjson-8     1920           1920           +0.00%
```

Here are the results for v7:

```
benchmark                                              old ns/op      new ns/op      delta
BenchmarkBulkDeleteRequestSerialization/stdlib-8       1149           1139           -0.87%
BenchmarkBulkDeleteRequestSerialization/easyjson-8     508            498            -1.97%
BenchmarkBulkIndexRequestSerialization/stdlib-8        2068           2087           +0.92%
BenchmarkBulkIndexRequestSerialization/easyjson-8      1458           1455           -0.21%
BenchmarkBulkEstimatedSizeInBytesWith1Request-8        1048930451     8359           -100.00%
BenchmarkBulkEstimatedSizeInBytesWith100Requests-8     877931         760386         -13.39%
BenchmarkBulkAllocs/1000_docs_with_64_byte-8           47375          31296          -33.94%
BenchmarkBulkAllocs/1000_docs_with_1_KiB-8             284910         177618         -37.66%
BenchmarkBulkAllocs/1000_docs_with_4_KiB-8             1161661        875848         -24.60%
BenchmarkBulkAllocs/1000_docs_with_16_KiB-8            5569934        3744312        -32.78%
BenchmarkBulkAllocs/1000_docs_with_64_KiB-8            15974479       12100076       -24.25%
BenchmarkBulkAllocs/1000_docs_with_256_KiB-8           84848153       63448152       -25.22%
BenchmarkBulkAllocs/1000_docs_with_1_MiB-8             3620659331     1607104063     -55.61%
BenchmarkBulkUpdateRequestSerialization/stdlib-8       2083           2160           +3.70%
BenchmarkBulkUpdateRequestSerialization/easyjson-8     1063           1011           -4.89%

benchmark                                              old allocs     new allocs     delta
BenchmarkBulkDeleteRequestSerialization/stdlib-8       9              9              +0.00%
BenchmarkBulkDeleteRequestSerialization/easyjson-8     5              5              +0.00%
BenchmarkBulkIndexRequestSerialization/stdlib-8        15             15             +0.00%
BenchmarkBulkIndexRequestSerialization/easyjson-8      11             11             +0.00%
BenchmarkBulkEstimatedSizeInBytesWith1Request-8        3255           47             -98.56%
BenchmarkBulkEstimatedSizeInBytesWith100Requests-8     4412           4412           +0.00%
BenchmarkBulkAllocs/1000_docs_with_64_byte-8           2              1              -50.00%
BenchmarkBulkAllocs/1000_docs_with_1_KiB-8             5              3              -40.00%
BenchmarkBulkAllocs/1000_docs_with_4_KiB-8             18             13             -27.78%
BenchmarkBulkAllocs/1000_docs_with_16_KiB-8            72             48             -33.33%
BenchmarkBulkAllocs/1000_docs_with_64_KiB-8            219            145            -33.79%
BenchmarkBulkAllocs/1000_docs_with_256_KiB-8           1254           939            -25.12%
BenchmarkBulkAllocs/1000_docs_with_1_MiB-8             15029          15028          -0.01%
BenchmarkBulkUpdateRequestSerialization/stdlib-8       19             19             +0.00%
BenchmarkBulkUpdateRequestSerialization/easyjson-8     10             10             +0.00%

benchmark                                              old bytes      new bytes      delta
BenchmarkBulkDeleteRequestSerialization/stdlib-8       1528           1528           +0.00%
BenchmarkBulkDeleteRequestSerialization/easyjson-8     1392           1392           +0.00%
BenchmarkBulkIndexRequestSerialization/stdlib-8        1144           1144           +0.00%
BenchmarkBulkIndexRequestSerialization/easyjson-8      1008           1008           +0.00%
BenchmarkBulkEstimatedSizeInBytesWith1Request-8        1063608        5273           -99.50%
BenchmarkBulkEstimatedSizeInBytesWith100Requests-8     532480         532427         -0.01%
BenchmarkBulkAllocs/1000_docs_with_64_byte-8           229428         114720         -50.00%
BenchmarkBulkAllocs/1000_docs_with_1_KiB-8             2147278        1073896        -49.99%
BenchmarkBulkAllocs/1000_docs_with_4_KiB-8             8306363        4157514        -49.95%
BenchmarkBulkAllocs/1000_docs_with_16_KiB-8            33120149       16603976       -49.87%
BenchmarkBulkAllocs/1000_docs_with_64_KiB-8            134276097      67645012       -49.62%
BenchmarkBulkAllocs/1000_docs_with_256_KiB-8           591417038      312466184      -47.17%
BenchmarkBulkAllocs/1000_docs_with_1_MiB-8             5263544936     4214919768     -19.92%
BenchmarkBulkUpdateRequestSerialization/stdlib-8       2224           2224           +0.00%
BenchmarkBulkUpdateRequestSerialization/easyjson-8     1888           1888           +0.00%
```

That's significant. Thanks for finding this.